pyproject.toml
[tool.poetry]
name = "cursdrf"
version = "0.1.0"
description = ""
authors = ["user <dolmatova3010@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flake8 = "^7.1.1"
black = "^24.10.0"
isort = "^6.0.0"
django = "^5.1.5"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
pillow = "^11.1.0"
python-dotenv = "^1.0.1"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.4.0"
coverage = "^7.6.10"
drf-yasg = "^1.21.8"
celery = "^5.4.0"
redis = "^5.2.1"
django-celery-beat = "^2.7.0"
django-cors-headers = "^4.6.0"
python-telegram-bot = "^21.10"
rest-framework-simplejwt = "^0.0.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
